:truck: rename: 마이페이지 Talk 관련 컴포넌트, components/Lists/MyPageLists 폴더 이동
