+++ client/src/components/Lists/MyPageLists/Talk/DetailTalk.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { LetterButtonColor } from '../Buttons/LetterButton';
+import { LetterButtonColor } from '../../../Buttons/LetterButton';
 
 function DetailTalk() {
 	return (
